FMWK-752-update-mockery - fix races
