+++ io/aerospike/xdr/tcp_server.go
@@ -19,10 +19,8 @@ import (
 	"crypto/tls"
 	"errors"
 	"fmt"
-	"io"
 	"log/slog"
 	"net"
-	"os"
 	"sync"
 	"sync/atomic"
 	"time"
@@ -33,7 +31,7 @@ import (
 
 const (
 	defaultAddress        = ":8080"
-	defaultTimeout        = 1 * time.Second
+	defaultTimeout        = 100 * time.Millisecond
 	defaultQueueSize      = 256
 	defaultMaxConnections = 4096
 )
@@ -88,8 +86,9 @@ type TCPServer struct {
 	// Fields for internal connection serving.
 	listener          net.Listener
 	activeConnections atomic.Int32
-	wg                sync.WaitGroup
-	cancel            context.CancelFunc
+	// Wait group to monitor handlers.
+	handlersWg sync.WaitGroup
+	cancel     context.CancelFunc
 
 	// Results will be sent here.
 	resultChan chan *models.ASBXToken
@@ -154,7 +153,7 @@ func (s *TCPServer) Stop() error {
 	if !s.isActive.CompareAndSwap(true, false) {
 		return fmt.Errorf("server is not active")
 	}
-
+	// Stop all routines by context cancelling.
 	s.cancel()
 
 	if s.listener != nil {
@@ -162,8 +161,9 @@ func (s *TCPServer) Stop() error {
 			s.logger.Warn("failed to close tcp server listener", slog.Any("error", err))
 		}
 	}
+	// Wait all handlers to stop.
+	s.handlersWg.Wait()
 
-	s.wg.Wait()
 	close(s.resultChan)
 
 	s.logger.Info("server shutdown complete")
@@ -213,6 +213,7 @@ func (s *TCPServer) acceptConnections(ctx context.Context) {
 			if s.activeConnections.Load() < int32(s.config.MaxConnections) {
 				// Increment connections counter.
 				s.activeConnections.Add(1)
+				s.handlersWg.Add(1)
 
 				go func() {
 					// Create and run handler.
@@ -230,6 +231,7 @@ func (s *TCPServer) acceptConnections(ctx context.Context) {
 					s.logger.Debug("connection finished",
 						slog.String("address", conn.RemoteAddr().String()))
 					s.activeConnections.Add(-1)
+					s.handlersWg.Done()
 				}()
 
 				s.logger.Debug("accepted new connection",
@@ -245,253 +247,4 @@ func (s *TCPServer) acceptConnections(ctx context.Context) {
 			}
 		}
 	}
-}
-
-// ConnectionHandler manages a single connection and its acknowledgment queue.
-type ConnectionHandler struct {
-	conn net.Conn
-	// channel to send results.
-	resultChan chan *models.ASBXToken
-	// Timeouts in nanoseconds.
-	readTimeoutNano  int64
-	writeTimeoutNano int64
-	// To stop all goroutines from inside.
-	cancel context.CancelFunc
-	wg     sync.WaitGroup
-	// Queue to process received messages.
-	bodyQueue chan []byte
-	// Queue to process ack messages.
-	ackQueue chan []byte
-
-	// Cached values.
-	ackMsgSuccess []byte
-	ackMsgRetry   []byte
-	timeNow       int64
-
-	logger           *slog.Logger
-	metricsCollector *metrics.Collector
-}
-
-// NewConnectionHandler returns a new connection handler.
-// A separate handler must be created for each connection.
-func NewConnectionHandler(
-	conn net.Conn,
-	resultChan chan *models.ASBXToken,
-	ackQueueSize int,
-	readTimeout time.Duration,
-	writeTimeout time.Duration,
-	logger *slog.Logger,
-	metricsCollector *metrics.Collector,
-) *ConnectionHandler {
-	return &ConnectionHandler{
-		conn:             conn,
-		resultChan:       resultChan,
-		readTimeoutNano:  readTimeout.Nanoseconds(),
-		writeTimeoutNano: writeTimeout.Nanoseconds(),
-		timeNow:          time.Now().UnixNano(),
-		bodyQueue:        make(chan []byte, ackQueueSize),
-		ackQueue:         make(chan []byte, ackQueueSize),
-		ackMsgSuccess:    NewAckMessage(AckOK),
-		ackMsgRetry:      NewAckMessage(AckRetry),
-		logger:           logger,
-		metricsCollector: metricsCollector,
-	}
-}
-
-// Start launches goroutines to serve the current connection.
-func (h *ConnectionHandler) Start(ctx context.Context) {
-	ctx, cancel := context.WithCancel(ctx)
-	h.cancel = cancel
-
-	h.wg.Add(4)
-
-	// This function serve h.timeNow field, to save each 100 milliseconds, current time.
-	// This time is used to update deadlines on write and read operations, to improve speed.
-	go func() {
-		defer h.wg.Done()
-
-		ticker := time.NewTicker(100 * time.Millisecond) // Update every 100ms
-		defer ticker.Stop()
-
-		for {
-			select {
-			case <-ctx.Done():
-				return
-			case t := <-ticker.C:
-				atomic.StoreInt64(&h.timeNow, t.UnixNano())
-			}
-		}
-	}()
-
-	go func() {
-		defer h.wg.Done()
-		h.handleMessages(ctx)
-		h.logger.Debug("message handler closed")
-	}()
-
-	go func() {
-		defer h.wg.Done()
-		h.processMessage(ctx)
-		h.logger.Debug("message processor closed")
-	}()
-
-	go func() {
-		defer h.wg.Done()
-		h.handleAcknowledgments(ctx)
-		h.logger.Debug("ack handler closed")
-		// After we gracefully closed all channels and stopped goroutines,
-		// we can cancel context to stop h.timeNow serving goroutine.
-		h.cancel()
-	}()
-
-	h.wg.Wait()
-	// When all routines finished we close the connection.
-	if err := h.conn.Close(); err != nil {
-		h.logger.Warn("failed to close connection", slog.Any("error", err))
-	}
-
-	h.logger.Debug("connection closed")
-}
-
-// handleMessages processes incoming messages.
-func (h *ConnectionHandler) handleMessages(ctx context.Context) {
-	parser := NewParser(h.conn)
-	// On exit from this function, we close this channel to send signal for the next goroutine to stop.
-	defer close(h.bodyQueue)
-
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		default:
-			deadline := h.getDeadline(h.readTimeoutNano)
-
-			if err := h.conn.SetReadDeadline(time.Unix(deadline, 0)); err != nil {
-				h.logger.Error("failed to set read deadline", slog.Any("error", err))
-
-				return
-			}
-
-			message, err := parser.Read()
-
-			switch {
-			case err == nil:
-				// ok.
-			case errors.Is(err, io.EOF):
-				// Exit if there is no message. Aerospike will open new connection if needed.
-				return
-			case os.IsTimeout(errors.Unwrap(err)):
-				// If timeout reached and the connection is closed, do nothing.
-				return
-			default:
-				h.logger.Error("failed to read message", slog.Any("error", err))
-				return
-			}
-
-			h.metricsCollector.Increment()
-
-			// Process message asynchronously
-			h.bodyQueue <- message
-		}
-	}
-}
-
-// processMessage serves h.bodyQueue. When a message is received, we try to parse it
-// and send it to h.resultChan, also ack messages is created for this message and sent to h.ackQueue.
-func (h *ConnectionHandler) processMessage(ctx context.Context) {
-	// On exit from this function, we close this channel to send signal for the next goroutine to stop.
-	defer close(h.ackQueue)
-
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case message, ok := <-h.bodyQueue:
-			if !ok {
-				return
-			}
-			// Parse message.
-			aMsg, err := ParseAerospikeMessage(message)
-			if err != nil {
-				h.logger.Error("failed to parse aerospike message", slog.Any("error", err))
-				// If we have an error on parsing message, we send an ack message with retry.
-				h.ackQueue <- h.ackMsgRetry
-
-				return
-			}
-			// Create aerospike key.
-			key, err := NewAerospikeKey(aMsg.Fields)
-
-			switch {
-			case err == nil:
-				// ok
-			case errors.Is(err, errSkipRecord):
-				// Send acknowledgement and skip record.
-				h.ackQueue <- h.ackMsgSuccess
-				continue
-			default:
-				h.logger.Error("failed to parse aerospike key", slog.Any("error", err))
-				// If we have an error on parsing message, we send an ack message with retry.
-				h.ackQueue <- h.ackMsgRetry
-
-				return
-			}
-			// Prepare payload.
-			// Reset xdr bit.
-			message = ResetXDRBit(message)
-			// Add headers.
-			payload := NewPayload(message)
-			// Create token ASBXToken.
-			token := models.NewASBXToken(key, payload)
-			// Send ASBXToken to results queue.
-			h.resultChan <- token
-
-			// Make acknowledgement.
-			h.ackQueue <- h.ackMsgSuccess
-		}
-	}
-}
-
-// handleAcknowledgments manages the async sending of XDR acks.
-// It receives messages from h.ackQueue and writes them to the connection socket.
-func (h *ConnectionHandler) handleAcknowledgments(ctx context.Context) {
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case ack, ok := <-h.ackQueue:
-			if !ok {
-				return
-			}
-			// Process each received ack message.
-			if err := h.sendAck(ack); err != nil {
-				h.logger.Warn("failed to send ack", slog.Any("error", err))
-				// Close connection!
-				h.cancel()
-
-				return
-			}
-		}
-	}
-}
-
-// sendAck updates connection deadline and writes an ack message to the connection.
-func (h *ConnectionHandler) sendAck(ack []byte) error {
-	deadline := h.getDeadline(h.writeTimeoutNano)
-	if err := h.conn.SetWriteDeadline(time.Unix(0, deadline)); err != nil {
-		return fmt.Errorf("failed to set ack write deadline: %w", err)
-	}
-
-	// Write an ack message to connection.
-	if _, err := h.conn.Write(ack); err != nil {
-		return fmt.Errorf("error writing ack to connection: %w", err)
-	}
-
-	return nil
-}
-
-// getDeadline loads current time from h.timeNow and adds timeout value.
-func (h *ConnectionHandler) getDeadline(timeout int64) int64 {
-	timeNow := atomic.LoadInt64(&h.timeNow)
-	return timeNow + timeout
 }
